Menu: extract the Face Display tab into build_face_display.cpp
Verbatim move of ProtoTracer (faces/colors/palette/animations), the
Protoface stack (face options, effects + layered builder, gradient
material, backend, HUB75 layout, size & position, panel preview) and
the boop/light/voice/LED drivers. The expression slot tables,
import_face_into_slot and the LayerCfg/LayeredEffectState structs move
along (Face Display only). The shared GIF preview machinery stays in
build_menu and is handed over via ctx.gif_leaf / ctx.draw_gif_preview.
